fix(core): update service deactivation to invoke missing handlers
